+++ src/main/webapp/app/fileupload/manage/assess/file-upload-assessment.component.ts
@@ -13,8 +13,8 @@ import { Complaint, ComplaintType } from 'app/entities/complaint.model';
 import { Course } from 'app/entities/course.model';
 import { ExerciseType, getCourseFromExercise } from 'app/entities/exercise.model';
 import { Feedback } from 'app/entities/feedback.model';
-import { FileUploadExercise } from 'app/entities/file-upload-exercise.model';
-import { FileUploadSubmission } from 'app/entities/file-upload-submission.model';
+import { FileUploadExercise } from 'app/fileupload/shared/entities/file-upload-exercise.model';
+import { FileUploadSubmission } from 'app/fileupload/shared/entities/file-upload-submission.model';
 import { StudentParticipation } from 'app/entities/participation/student-participation.model';
 import { Result } from 'app/entities/result.model';
 import { getLatestSubmissionResult, getSubmissionResultById } from 'app/entities/submission.model';
