+++ src/main/webapp/app/fileupload/shared/entities/file-upload-exercise.model.ts
