+++ src/main/webapp/app/fileupload/shared/entities/file-upload-submission.model.ts
