+++ src/test/javascript/spec/component/assessment-dashboard/assessment-dashboard.component.spec.ts
@@ -9,7 +9,7 @@ import { TutorParticipationStatus } from 'app/entities/participation/tutor-parti
 import { ExerciseService } from 'app/exercise/exercise.service';
 import { StatsForDashboard } from 'app/course/dashboards/stats-for-dashboard.model';
 import { TextExercise } from 'app/entities/text/text-exercise.model';
-import { FileUploadExercise } from 'app/entities/file-upload-exercise.model';
+import { FileUploadExercise } from 'app/fileupload/shared/entities/file-upload-exercise.model';
 import { ProgrammingExercise } from 'app/entities/programming/programming-exercise.model';
 import { Exam } from 'app/entities/exam/exam.model';
 import { ExerciseGroup } from 'app/entities/exercise-group.model';
