+++ src/test/javascript/spec/component/assessment-shared/assessment-instructions.component.spec.ts
@@ -6,7 +6,7 @@ import { StructuredGradingInstructionsAssessmentLayoutComponent } from 'app/asse
 import { ModelingExercise } from 'app/entities/modeling-exercise.model';
 import { ArtemisMarkdownService } from 'app/shared/markdown.service';
 import { TextExercise } from 'app/entities/text/text-exercise.model';
-import { FileUploadExercise } from 'app/entities/file-upload-exercise.model';
+import { FileUploadExercise } from 'app/fileupload/shared/entities/file-upload-exercise.model';
 import { ProgrammingExercise } from 'app/entities/programming/programming-exercise.model';
 import { ExerciseType } from 'app/entities/exercise.model';
 import { ModelingEditorComponent } from 'app/modeling/shared/modeling-editor.component';
