+++ src/test/javascript/spec/component/assessment-shared/assessment-locks.component.spec.ts
@@ -7,7 +7,7 @@ import { MockRouterLinkDirective } from '../../helpers/mocks/directive/mock-rout
 import { MockHasAnyAuthorityDirective } from '../../helpers/mocks/directive/mock-has-any-authority.directive';
 import { ArtemisDatePipe } from 'app/shared/pipes/artemis-date.pipe';
 import { ModelingSubmission } from 'app/entities/modeling-submission.model';
-import { FileUploadSubmission } from 'app/entities/file-upload-submission.model';
+import { FileUploadSubmission } from 'app/fileupload/shared/entities/file-upload-submission.model';
 import { TextSubmission } from 'app/entities/text/text-submission.model';
 import { ProgrammingSubmission } from 'app/entities/programming/programming-submission.model';
 import { ModelingAssessmentService } from 'app/modeling/manage/assess/modeling-assessment.service';
