+++ src/test/javascript/spec/component/file-upload-exercise/file-upload-exercise-update.component.spec.ts
@@ -5,7 +5,7 @@ import { of, Subject } from 'rxjs';
 
 import { FileUploadExerciseUpdateComponent } from 'app/fileupload/manage/file-upload-exercise-update.component';
 import { FileUploadExerciseService } from 'app/fileupload/manage/file-upload-exercise.service';
-import { FileUploadExercise } from 'app/entities/file-upload-exercise.model';
+import { FileUploadExercise } from 'app/fileupload/shared/entities/file-upload-exercise.model';
 import { MockSyncStorage } from '../../helpers/mocks/service/mock-sync-storage.service';
 import { LocalStorageService, SessionStorageService } from 'ngx-webstorage';
 import { MockActivatedRoute } from '../../helpers/mocks/activated-route/mock-activated-route';
