+++ src/test/javascript/spec/helpers/mocks/service/mock-file-upload-submission.service.ts
@@ -1,7 +1,7 @@
 import { of } from 'rxjs';
 import { fileUploadExercise } from './mock-file-upload-exercise.service';
 import { Result } from 'app/entities/result.model';
-import { FileUploadSubmission } from 'app/entities/file-upload-submission.model';
+import { FileUploadSubmission } from 'app/fileupload/shared/entities/file-upload-submission.model';
 import { StudentParticipation } from 'app/entities/participation/student-participation.model';
 
 export const fileUploadParticipation = new StudentParticipation();
