+++ src/test/playwright/e2e/exercise/file-upload/FileUploadExerciseAssessment.spec.ts
@@ -1,5 +1,5 @@
 import { Course } from 'app/entities/course.model';
-import { FileUploadExercise } from 'app/entities/file-upload-exercise.model';
+import { FileUploadExercise } from 'app/fileupload/shared/entities/file-upload-exercise.model';
 
 import { admin, instructor, studentOne, tutor } from '../../../support/users';
 import { test } from '../../../support/fixtures';
