+++ src/test/playwright/e2e/exercise/file-upload/FileUploadExerciseManagement.spec.ts
@@ -1,7 +1,7 @@
 import dayjs from 'dayjs';
 
 import { Course } from 'app/entities/course.model';
-import { FileUploadExercise } from 'app/entities/file-upload-exercise.model';
+import { FileUploadExercise } from 'app/fileupload/shared/entities/file-upload-exercise.model';
 
 import { admin } from '../../../support/users';
 import { generateUUID } from '../../../support/utils';
