+++ src/test/playwright/e2e/exercise/file-upload/FileUploadExerciseParticipation.spec.ts
@@ -1,10 +1,10 @@
 import { Course } from 'app/entities/course.model';
-import { FileUploadExercise } from 'app/entities/file-upload-exercise.model';
+import { FileUploadExercise } from 'app/fileupload/shared/entities/file-upload-exercise.model';
 
 import { admin, studentOne } from '../../../support/users';
 import { test } from '../../../support/fixtures';
 import { expect } from '@playwright/test';
-import { FileUploadSubmission } from 'app/entities/file-upload-submission.model';
+import { FileUploadSubmission } from 'app/fileupload/shared/entities/file-upload-submission.model';
 
 test.describe('File upload exercise participation', { tag: '@fast' }, () => {
     let course: Course;
