+++ src/test/playwright/support/requests/ExerciseAPIRequests.ts
@@ -32,7 +32,7 @@ import {
 import { dayjsToString, generateUUID, titleLowercase } from '../utils';
 import { ModelingExercise } from 'app/entities/modeling-exercise.model';
 import { ProgrammingExercise } from 'app/entities/programming/programming-exercise.model';
-import { FileUploadExercise } from 'app/entities/file-upload-exercise.model';
+import { FileUploadExercise } from 'app/fileupload/shared/entities/file-upload-exercise.model';
 import { Participation } from 'app/entities/participation/participation.model';
 import { Exam } from 'app/entities/exam/exam.model';
 import { StudentParticipation } from 'app/entities/participation/student-participation.model';
